LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointerrawupdate-expected.txt


PASS pointerrawupdate event is not fired
PASS onpointerrawupdate is not exposed
